% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{hltsim}
\alias{hltsim}
\title{Simulate the HLT model}
\usage{
hltsim(type, n, ntheta, lambda, id, dL, nB, beta = NULL)
}
\arguments{
\item{n}{number of observations}

\item{ntheta}{number first-level of latent dimensions}

\item{lambda}{latent factor coefficients}

\item{id}{number of questions}

\item{dL}{number of levels for each question}

\item{nB}{number of regression parameters. nB = ncol(z).}

\item{beta}{what value to set the regression parameters.}

\item{mua}{mean of the alphas}

\item{mud}{mean of the deltas}

\item{siga}{standard deviation of the alphas}

\item{sigd}{standard deviation of the deltas}

\item{regression}{TRUE/FALSE. Simulate regression model?}

\item{z}{regression design matrix.}
}
\description{
Simulate the HLT model
}
\examples{
# PCM, No regression, 2 dimensions, binary response
xdat = hltsim(type = "1p", n = 50, ntheta = 2, lambda = c(0.5, 0.5), 
              id = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1), dL = 2)
apply(xdat$x, 2, table)
mod = hlt(xdat$x, id = xdat$id, iter = 5e5, burn = 3e5, type = "1p", 
          delta = 0.1)
          
# PCM, No regression, 3 dimensions, binary response
xdat = hltsim(type = "1p", n = 200, ntheta = 3, lambda = c(0.2, 0.2, 0.1), 
              id = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1,2,2,2,2,2,2,2,2,2,2), dL = 2)
apply(xdat$x, 2, table)
mod = hlt(xdat$x, id = xdat$id, iter = 5e5, burn = 3e5, type = "1p", 
          delta = 0.04)

# PCM, No regression, 2 dimensions
xdat = hltsim(n = n, ntheta = 2, lambda = c(0.7, 0.7), 
              tJ = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1),
              dL = 10, mua = NULL, mud = 1.4, siga = NULL, sigd = 1,
              regression = FALSE, z = NULL)
apply(xdat$x, 2, table)
mod = hlt(xdat$x, id = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1), 
          iter = 200, burn = 100, delta = 0.013)

# PCM, No regression, 4 dimensions
type, n, ntheta, lambda, id, dL, nB, beta = NULL
xdat = hltsim(type = "1p", n = 50, ntheta = 2, lambda = c(0.5, 0.5), 
              id = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1), dL = 2)
apply(xdat$x, 2, table)
mod = hlt(xdat$x, id = xdat$id, iter = 5e5, burn = 3e5, delta = 0.03)


# GPCM, No regression, 2 dimensions
xdat = hltsim(n = n, ntheta = 2, lambda = c(0.7, 0.7), 
              tJ = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1),
              dL = 10, mua = NULL, mud = 1.4, siga = NULL, sigd = 1,
              regression = FALSE, z = NULL)
apply(xdat$x, 2, table)
mod = hlt(xdat$x, id = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1), 
          iter = 200, burn = 100, delta = 0.013)

# GPCM, Regression, 2 dimensions
xdat = hltsim(n = n, ntheta = 2, lambda = c(0.7, 0.7), 
              tJ = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1),
              dL = 10, mua = NULL, mud = 1.4, siga = NULL, sigd = 1,
              regression = FALSE, z = z)
apply(xdat$x, 2, table)
mod = hlt(xdat$x, z = z, id = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1), 
          iter = 200, burn = 100, delta = 0.013)

# GPCM, No regression, 4 dimensions
xdat = hltsim(n = n, ntheta = 4, lambda = c(0.7, 0.2, 1.2, 0.4), 
              tJ = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1,2,2,2,2,2,2, 
                     3,3,3,3,3,3),
              dL = 10, mua = NULL, mud = 1.4, siga = NULL, sigd = 1,
              regression = FALSE, z = NULL)
apply(xdat$x, 2, table)
mod = hlt(xdat$x, id = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1,2,2,2,2,2,2, 
          3,3,3,3,3,3), iter = 200, burn = 100, delta = 0.013)

# GPCM, Regression, 4 dimensions
xdat = hltsim(n = n, ntheta = 4, lambda = c(0.7, 0.2, 1.2, 0.4), 
              tJ = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1,2,2,2,2,2,2, 
                     3,3,3,3,3,3),
              dL = 10, mua = NULL, mud = 1.4, siga = NULL, sigd = 1,
              regression = FALSE, z = z)
apply(xdat$x, 2, table)
mod = hlt(xdat$x, z = z, id = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1,2,2,2,2,2,2, 
          3,3,3,3,3,3), iter = 200, burn = 100, delta = 0.013)




}
