% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html}
\alias{html}
\title{html}
\usage{
html(x, pretty = FALSE, level = 0, indent = "    ", ...)
}
\arguments{
\item{x}{An object to be coerced to html}

\item{pretty}{Fancy indentation?}

\item{level}{Indentation level}

\item{indent}{String for indentation}

\item{...}{Other parameters}
}
\description{
html
}
