% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_body}
\alias{html_body}
\title{Create a \sQuote{<body>} tag in an \sQuote{HTML} page}
\usage{
html_body(x)
}
\arguments{
\item{x}{the content to display.}
}
\value{
an \code{html_object}.
}
\description{
Create a \sQuote{<body>} tag in an \sQuote{HTML} page
}
