% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_comment}
\alias{html_comment}
\title{Create a \sQuote{<comment>} tag}
\usage{
html_comment(content, ...)
}
\arguments{
\item{...}{tag attributes.}
}
\value{
an \code{html_object}.
}
\description{
Create a \sQuote{<comment>} tag
}
