% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_div}
\alias{html_div}
\title{Create a \sQuote{<div>} tag}
\usage{
html_div(content, ...)
}
\arguments{
\item{content}{of the div tag.}

\item{...}{tag attributes.}
}
\value{
an \code{html_object}.
}
\description{
Create a \sQuote{<div>} tag
}
