% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_foot}
\alias{html_foot}
\title{Create a \sQuote{<tfoot>} tag in an \sQuote{HTML} table}
\usage{
html_foot(x)
}
\arguments{
\item{x}{one row \code{data.frame} with the content of the foot.}
}
\value{
an \code{html_object}.
}
\description{
Create a \sQuote{<tfoot>} tag in an \sQuote{HTML} table
}
