% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_H2}
\alias{html_H2}
\title{Create a \sQuote{H2} tag}
\usage{
html_H2(x, ...)
}
\arguments{
\item{x}{the content to display.}

\item{...}{tag attributes.}
}
\value{
an \code{html_object}.
}
\description{
Create a \sQuote{H2} tag
}
