% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_H4}
\alias{html_H4}
\title{Create a \sQuote{H4} tag}
\usage{
html_H4(x, ...)
}
\arguments{
\item{x}{the content to display.}

\item{...}{tag attributes.}
}
\value{
an \code{html_object}.
}
\description{
Create a \sQuote{H4} tag
}
