% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_html}
\alias{html_html}
\title{Create a \sQuote{<html>} tag in an \sQuote{HTML} page}
\usage{
html_html(x, ...)
}
\arguments{
\item{x}{the content to display.}

\item{...}{tag attributes.}
}
\value{
an \code{html_object}.
}
\description{
Create a \sQuote{<html>} tag in an \sQuote{HTML} page
}
