% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_img}
\alias{html_img}
\title{Create a \sQuote{img} tag in an \sQuote{HTML} page}
\usage{
html_img(src, alt, ...)
}
\arguments{
\item{src}{URL of the image.}

\item{alt}{alternate text for the image.}

\item{...}{tag attributes.}
}
\value{
an \code{html_object}.
}
\description{
Create a \sQuote{img} tag in an \sQuote{HTML} page
}
