% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_link}
\alias{html_link}
\title{Create a \sQuote{<link>} tag}
\usage{
html_link(rel, ...)
}
\arguments{
\item{rel}{relationship between document and linked document.}

\item{...}{tag attributes.}
}
\value{
an \code{html_object}.
}
\description{
Create a \sQuote{<link>} tag
}
