% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_meta}
\alias{html_meta}
\title{Create a \sQuote{<meta>} tag}
\usage{
html_meta(...)
}
\arguments{
\item{...}{tag attributes.}
}
\value{
an \code{html_object}.
}
\description{
Create a \sQuote{<meta>} tag
}
\examples{
\dontrun{
library(hlt)
html_meta(charset = "utf-8")
html_meta("http-equiv" = "x-ua-compatible")
}
}
