% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_object}
\alias{html_object}
\title{html_object}
\usage{
html_object(.tag, .content = NULL, ...)
}
\arguments{
\item{.tag}{The tag of the html object}

\item{.content}{The content of the html object}

\item{...}{Other arguments}
}
\description{
html_object
}
