% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_script}
\alias{html_script}
\title{Create a \sQuote{<script>} tag}
\usage{
html_script(content, ...)
}
\arguments{
\item{content}{of the script tag.}

\item{...}{tag attributes.}
}
\value{
an \code{html_object}.
}
\description{
Create a \sQuote{<script>} tag
}
