% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_style}
\alias{html_style}
\title{Create a \sQuote{<style>} tag}
\usage{
html_style(content, ...)
}
\arguments{
\item{content}{style information for an HTML document.}

\item{...}{tag attributes.}
}
\value{
an \code{html_object}.
}
\description{
Create a \sQuote{<style>} tag
}
