% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_table}
\alias{html_table}
\title{Create an \sQuote{HTML} table}
\usage{
html_table(x)
}
\arguments{
\item{x}{\code{data.frame} with the content of the table.}
}
\value{
an \code{html_object}.
}
\description{
Create an \sQuote{HTML} table
}
