% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_tbody}
\alias{html_tbody}
\title{Create a \sQuote{<tbody>} tag in an \sQuote{HTML} table}
\usage{
html_tbody(x)
}
\arguments{
\item{x}{one row \code{data.frame} with the content of the row.}
}
\value{
an \code{html_object}.
}
\description{
Create a \sQuote{<tbody>} tag in an \sQuote{HTML} table
}
