% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_td}
\alias{html_td}
\title{Create a cell in an \sQuote{HTML} table}
\usage{
html_td(x, ...)
}
\arguments{
\item{x}{the content to display.}

\item{...}{tag attributes.}
}
\value{
an \code{html_object}.
}
\description{
Create a cell in an \sQuote{HTML} table
}
