% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_th}
\alias{html_th}
\title{Create a header cell in an \sQuote{HTML} table}
\usage{
html_th(x)
}
\arguments{
\item{x}{the content to display.}
}
\value{
an \code{html_object}.
}
\description{
Create a header cell in an \sQuote{HTML} table
}
