% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_title}
\alias{html_title}
\title{Create a \sQuote{<title>} tag}
\usage{
html_title(title, ...)
}
\arguments{
\item{title}{the title of the page.}

\item{...}{tag attributes.}
}
\value{
an \code{html_object}.
}
\description{
Create a \sQuote{<title>} tag
}
