% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html_tr}
\alias{html_tr}
\title{Create a row in an \sQuote{HTML} table}
\usage{
html_tr(x)
}
\arguments{
\item{x}{one row \code{data.frame} with the content of the row.}
}
\value{
an \code{html_object}.
}
\description{
Create a row in an \sQuote{HTML} table
}
