% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R, R/merge_chains.R
\name{merge_chains}
\alias{merge_chains}
\title{Merge Chains from hlt method}
\usage{
merge_chains(x, ...)

merge_chains(x, ...)
}
\arguments{
\item{x}{object of class "hltObjList"}

\item{...}{other arguments}
}
\description{
Merge Chains from hlt method

Merge Chains from hlt method
}
