% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{tag_attr}
\alias{tag_attr}
\title{Add attributes to a tag}
\usage{
tag_attr(tag, ...)
}
\arguments{
\item{tag}{add attributes to tag.}

\item{...}{named attributes to add e.g. \code{style =
"background-color: lightblue;"}.}
}
\value{
tag
}
\description{
Add attributes to a tag
}
\examples{
\dontrun{
h <- html_html(html_body(
    html_p("Display the 'cars' dataset as a table") +
    html_table(cars) +
    tag_attr(html_p("Disaplay the 'cars' dataset again, but now with column sums in a 'tfoot' tag"),
             style = "background-color: lightblue;") +
    html_table(rbind(cars, colSums(cars)), tfoot = TRUE) +
    html_p("Display the 'mtcars' dataset as a table") +
    html_table(mtcars) +
    html_p("Nice &#9786;")))

filename <- tempfile(fileext = ".html")
capture.output(h, file = filename)
browseURL(filename)
}
}
