% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{td}
\alias{td}
\title{Create a standard cell in an \sQuote{HTML} table}
\usage{
td(x, style = NULL)
}
\arguments{
\item{x}{the content to display.}

\item{style}{the style to apply.}
}
\value{
an \code{html_td} object.
}
\description{
Create a standard cell in an \sQuote{HTML} table
}
