% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{hltsim}
\alias{hltsim}
\title{Simulate the HLT model}
\usage{
hltsim(
  n,
  ntheta,
  lambda,
  tJ,
  dL,
  mua,
  mud,
  siga,
  sigd,
  regression = FALSE,
  z = NULL,
  nB,
  beta = NULL
)
}
\arguments{
\item{n}{number of observations}

\item{ntheta}{number first-level of latent dimensions}

\item{lambda}{latent factor coefficients}

\item{tJ}{number of questions}

\item{dL}{number of levels for each question}

\item{mua}{mean of the alphas}

\item{mud}{mean of the deltas}

\item{siga}{standard deviation of the alphas}

\item{sigd}{standard deviation of the deltas}
}
\description{
Simulate the HLT model
}
\examples{
xdat = hltsim(n = 100, ntheta = 3, lambda = c(0.8, 0.4, 0.3), 
              tJ = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1,2,2,2,2,2,2,2,2,2),
              dL = 5, mua = 1, mud = 1.4, siga = 0.8, sigd = 1,
              regression = FALSE, z = NULL)
apply(xdat$x, 2, table)

nB = 1
n = 100
z = matrix(sample(0:1, n, replace = TRUE), nrow = n, ncol = nB)
xdat = hltsim(n = 100, ntheta = 3, lambda = c(0.8, 0.4, 0.3), 
              tJ = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1,2,2,2,2,2,2,2,2,2),
              dL = 5, mua = 1, mud = 1.4, siga = 0.8, sigd = 1,
              regression = TRUE, z = z, nB = nB)
apply(xdat$x, 2, table)
lm(xdat$theta[,4] ~ xdat$z)
xdat$s.beta


# 2 dimension examples

# PCM

xdat = hltsim(n = 100, ntheta = 2, lambda = c(0.7, 0.7), 
              tJ = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1),
              dL = 2, mua = NULL, mud = 1.4, siga = NULL, sigd = 1,
              regression = FALSE, z = NULL)
apply(xdat$x, 2, table)

nB = 5
z = matrix(sample(0:1, n, replace = TRUE), nrow = n, ncol = nB)
xdat = hltsim(n = 100, ntheta = 2, lambda = c(0.7, 0.7), 
              tJ = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1),
              dL = 2, mua = NULL, mud = 1.4, siga = NULL, sigd = 1,
              regression = TRUE, z = z)
apply(xdat$x, 2, table)


# GPCM

xdat = hltsim(n = 100, ntheta = 2, lambda = c(0.7, 0.7), 
              tJ = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1),
              dL = 7, mua = 1, mud = 1.4, siga = 0.8, sigd = 1,
              regression = FALSE, z = NULL)
apply(xdat$x, 2, table)

nB = 5
z = matrix(sample(0:1, n, replace = TRUE), nrow = n, ncol = nB)
xdat = hltsim(n = 100, ntheta = 2, lambda = c(0.7, 0.7), 
              tJ = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1),
              dL = 7, mua = 1, mud = 1.4, siga = 0.8, sigd = 1,
              regression = TRUE, z = z)
apply(xdat$x, 2, table)


# 4 dimension examples

# PCM

xdat = hltsim(n = 100, ntheta = 4, lambda = c(0.7, 0.2, 0.8, 0.4), 
              tJ = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1),
              dL = 2, mua = NULL, mud = 1.4, siga = NULL, sigd = 1,
              regression = FALSE, z = NULL)
apply(xdat$x, 2, table)

nB = 5
z = matrix(sample(0:1, n, replace = TRUE), nrow = n, ncol = nB)
xdat = hltsim(n = 100, ntheta = 4, lambda = c(0.7, 0.2, 0.8, 0.4), 
              tJ = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1),
              dL = 2, mua = NULL, mud = 1.4, siga = NULL, sigd = 1,
              regression = TRUE, z = z)
apply(xdat$x, 2, table)


# GPCM

xdat = hltsim(n = 100, ntheta = 4, lambda = c(0.7, 0.2, 0.8, 0.4), 
              tJ = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1),
              dL = 7, mua = 1, mud = 1.4, siga = 0.8, sigd = 1,
              regression = FALSE, z = NULL)
apply(xdat$x, 2, table)

nB = 5
z = matrix(sample(0:1, n, replace = TRUE), nrow = n, ncol = nB)
xdat = hltsim(n = 100, ntheta = 4, lambda = c(0.7, 0.2, 0.8, 0.4), 
              tJ = c(0,0,0,0,0,0,0,1,1,1,1,1,1,1,1,1,1),
              dL = 7, mua = 1, mud = 1.4, siga = 0.8, sigd = 1,
              regression = TRUE, z = z)
apply(xdat$x, 2, table)

}
