% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{hltsim}
\alias{hltsim}
\title{Simulate the HLT model}
\usage{
hltsim(
  n,
  ntheta,
  lambda,
  tJ,
  dL,
  mua,
  mud,
  siga,
  sigd,
  regression = FALSE,
  z = NULL,
  nB,
  beta = NULL
)
}
\arguments{
\item{n}{number of observations}

\item{ntheta}{number first-level of latent dimensions}

\item{lambda}{latent factor coefficients}

\item{tJ}{number of questions}

\item{dL}{number of levels for each question}

\item{mua}{mean of the alphas}

\item{mud}{mean of the deltas}

\item{siga}{standard deviation of the alphas}

\item{sigd}{standard deviation of the deltas}

\item{regression}{TRUE/FALSE. Simulate regression model?}

\item{z}{regression design matrix.}

\item{nB}{number of regression parameters. nB = ncol(z).}

\item{beta}{what value to set the regression parameters.}
}
\description{
Simulate the HLT model
}
